Function transaction_builder_generator::java::write_source_files

pub fn write_source_files(
    install_dir: PathBuf, 
    package_name: Option<&str>, 
    abis: &[ScriptABI]
) -> Result<()>

Output transaction builders and decoders in Java for the given ABIs. Source files will be located in a subdirectory of install_dir corresponding to the given package name (if any, otherwise install_dir it self).